kartya componentizálva
